+++ gear_optimizer/solver/inflight_utils.py
@@ -97,9 +97,9 @@ def _build_calc_song_from_file(*, fp: str, found_song_name: str, cfg, cfg_dict: 
         timestamps_raw = song_data.get("timestamps")
         note_types_raw = song_data.get("note_types")
         if isinstance(timestamps_raw, np.ndarray):
-            song_timestamps_np = timestamps_raw.astype(np.float64, copy=False)
+            song_timestamps_np = timestamps_raw.astype(np.float32, copy=False)
         else:
-            song_timestamps_np = np.array(timestamps_raw if timestamps_raw is not None else [], dtype=np.float64)
+            song_timestamps_np = np.array(timestamps_raw if timestamps_raw is not None else [], dtype=np.float32)
         if isinstance(note_types_raw, np.ndarray):
             song_note_types_np = note_types_raw.astype(np.int16, copy=False)
         else:
@@ -111,7 +111,7 @@ def _build_calc_song_from_file(*, fp: str, found_song_name: str, cfg, cfg_dict: 
         cached = _song_file_cache_get(fp)
     if cached is None:
         # Defensive fallback (shouldn't happen).
-        cached = ({}, np.array([], dtype=np.float64), np.array([], dtype=np.int16))
+        cached = ({}, np.array([], dtype=np.float32), np.array([], dtype=np.int16))
     meta0, song_timestamps_np, song_note_types_np = cached
 
     calc_song = {
